% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptmc_discrete.R
\name{get_discrete_output}
\alias{get_discrete_output}
\title{Get Discrete Output from PTMC Model}
\usage{
get_discrete_output(model, data_list, settings, update_ind, par)
}
\arguments{
\item{model}{A list representing the model containing information such as parameter names and any
model-specific settings required for running the PTMC.}

\item{data_list}{A list containing the data necessary for running the model. This could include
observed data, priors, and any other variables needed for the simulation.}

\item{settings}{A list containing the settings for the model execution. The list must contain:
- \code{numberChainRuns}: The number of chains to run in parallel.
- \code{runParallel}: A boolean indicating whether to run the chains in parallel (TRUE)
or sequentially (FALSE).
- \code{numberFittedPar}: The number of parameters to fit in the model.}

\item{update_ind}{An index or flag used to control which part of the model or data to update
during the PTMC simulation.}

\item{par}{A list of parameters or starting values for the chains. Each element corresponds to
one chain.}
}
\value{
A list containing the results of the PTMC simulation. The list includes:
\item{mcmc}{An MCMC object containing the posterior samples of the model parameters.}
\item{discrete}{A list of discrete states generated during the simulation for each chain.}
\item{lpost}{A data frame of log-posterior values, with columns representing different chains
and rows representing samples.}
\item{temp}{A data frame of temperatures for each chain at each sample.}
\item{acc}{A data frame of acceptance rates for each chain at each sample.}
\item{outPTpar}{A list containing the parameter values for each chain.}
}
\description{
This function runs multiple Markov Chains in parallel or sequentially to generate output from a
Parallel Tempering Monte Carlo (PTMC) model. The output includes the posterior parameter samples,
discrete state information, log posterior values, temperatures, and acceptance rates.
}
