% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptmc.R
\name{ptmc_func}
\alias{ptmc_func}
\title{Create a model}
\usage{
ptmc_func(model, data, settings)
}
\arguments{
\item{model}{The model to run ptmc for (build using \code{\link{build_model}}).}

\item{data}{Data used in the calibration process}

\item{settings}{settings}
}
\value{

}
\description{
Create a model
}
