\name{ptmc-package}
\alias{ptmc-package}
\alias{ptmc}
\docType{package}
\title{
\packageTitle{ptmc}
}
\description{
\packageDescription{ptmc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ptmc}
\packageIndices{ptmc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ptmc}

Maintainer: \packageMaintainer{ptmc}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
